Refactor code, move actions inside components
