+++ .eslintrc
@@ -20,6 +20,7 @@
     "import/newline-after-import": "error",
     "import/no-duplicates": "error",
     "import/no-named-default": "error",
+    "import/prefer-default-export": "off",
     "key-spacing": "error",
     "keyword-spacing": ["error", { "before": true, "after": true }],
     "no-console": "off",
