--- src/client/_actions/actionTypes.js
@@ -1,2 +1,0 @@
-export const FETCH_ITEMS = 'FETCH_ITEMS';
-export const ADD_ITEM = 'ADD_ITEM';
