--- src/client/_actions/itemActions.js
@@ -1,5 +1,0 @@
-import { ADD_ITEM } from './actionTypes';
-
-export default function addItem(item) {
-  return { type: ADD_ITEM, item };
-}
