--- src/client/_dispatchers/addItem.js
@@ -1,7 +1,0 @@
-import add from '../_actions/itemActions';
-
-const addItem = item => dispatch => {
-  dispatch(add(item));
-};
-
-export default addItem;
