+++ src/client/components/ProductsList/index.js
@@ -1,5 +1,3 @@
 import ProductsList from './ProductsList';
 
-export default {
-  ProductsList
-};
+export { ProductsList };
