+++ src/client/components/SearchBar/index.js
@@ -1,0 +1,3 @@
+import SearchBar from './SearchBar';
+
+export { SearchBar };
