+++ src/client/store/configureStore.js
@@ -2,7 +2,7 @@ import { createStore, applyMiddleware } from 'redux';
 import { composeWithDevTools } from 'redux-devtools-extension';
 import thunk from 'redux-thunk';
 
-import rootReducer from '../_reducers/rootReducer';
+import rootReducer from '../reducers/rootReducer';
 
 export default function configureStore() {
   return createStore(rootReducer, composeWithDevTools(applyMiddleware(thunk)));
